// web/src/types/wedmStudio.ts
export interface Point2D {
  x: number;
  y: number;
}

export interface LineSegment {
  type: "line";
  start: Point2D;
  end: Point2D;
}

export interface ArcSegment {
  type: "arc";
  start: Point2D;
  end: Point2D;
  center: Point2D;
  radius: number;
  start_angle_rad: number;
  end_angle_rad: number;
  ccw: boolean;
}

export type GeometrySegment = LineSegment | ArcSegment;

export interface ProfileContour {
  id: string;
  segments: GeometrySegment[];
  is_closed: boolean;
  is_exterior: boolean;
  area_mm2: number;
  perimeter_mm: number;
  bbox: { min_x: number; min_y: number; max_x: number; max_y: number };
}